move useItems up a level
